refactor(proof): split optimism rpc utility
