+++ crates/utilities/optimism-rpc/README.md
@@ -1,0 +1,3 @@
+# base-optimism-rpc
+
+Typed Optimism rollup JSON-RPC provider extensions and response types.
